// fake-node/service/src/lib/common/types/bet-status.ts
export type BetStatus =
  | 'BET' // 이것은 시작 베팅이며 플레이어 베팅이 확인되었지만 아직 정산되지 않았습니다. 게임 결과 및 정산을 기다리고 있습니다.
  | 'TIP' // 스트리머에게 팁을 준 상태입니다.
  | 'GIVE' // 게임사에서 프로모션 돈을 받은 상태입니다.
  | 'JACKPOT' // 잭팟 상태입니다.
  | 'BONUS' // 게임 스핀중에 보너스
  | 'ADJUST' // 잭팟 상태입니다.
  | 'SETTLE' // 성공적인 베팅에 대한 게임 결과를 업데이트하고 플레이어 잔액을 업데이트하십시오.
  | 'VOID' // 어떤 이유(치팅, 게임 오류, 작업 오류 등)로 인해 성공적으로 생성된 트랜잭션을 무효화해야 합니다.
  | 'CANCEL' // 데이터가 동기화되지 않거나 시간 초과되어 운영자와 게임 제공자가 올바르게 응답하지 않으면 시스템에서 베팅 취소 조치를 요청합니다.